% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{.available_filters_choices.db}
\alias{.available_filters_choices.db}
\title{Generate available filters choices based on the db Source data}
\usage{
\method{.available_filters_choices}{db}(source, cohort, ...)
}
\arguments{
\item{source}{Source object of db type.}

\item{cohort}{cohortBuilder cohort object.}

\item{...}{Extra arguments passed to a specific method.}
}
\value{
`shinyWidgets::prepare_choices` output value.
}
\description{
The method returns the available choices for the virtualSelect input.
}
