% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_date_range.R
\name{cb_filter.date_range.db}
\alias{cb_filter.date_range.db}
\title{Date range filter definition}
\usage{
\method{cb_filter.date_range}{db}(
  source,
  type = "date_range",
  id = .gen_id(),
  name = id,
  variable,
  range = NA,
  dataset,
  keep_na = TRUE,
  ...,
  description = NULL,
  active = TRUE
)
}
\description{
This is a method for cohortBuilder [cb_filter.date_range()] generic.
It creates 'date_range' filter type definition for specified table database connection.
}
