% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_discrete.R
\name{cb_filter.discrete.db}
\alias{cb_filter.discrete.db}
\title{Discrete filter definition}
\usage{
\method{cb_filter.discrete}{db}(
  source,
  type = "discrete",
  id = .gen_id(),
  name = id,
  variable,
  value = NA,
  dataset,
  keep_na = TRUE,
  ...,
  description = NULL,
  active = TRUE
)
}
\description{
This is a method for cohortBuilder [cb_filter.discrete()] generic.
It creates 'discrete' filter type definition for specified table database connection.
}
