% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_range.R
\name{cb_filter.range.db}
\alias{cb_filter.range.db}
\title{Range filter definition}
\usage{
\method{cb_filter.range}{db}(
  source,
  type = "range",
  id = .gen_id(),
  name = id,
  variable,
  range = NA,
  dataset,
  keep_na = TRUE,
  ...,
  description = NULL,
  active = TRUE
)
}
\description{
This is a method for cohortBuilder [cb_filter.range()] generic.
It creates 'range' filter type definition for specified table database connection.
}
