% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_discrete_text.R
\name{db-filter-discrete-text}
\alias{db-filter-discrete-text}
\title{Discrete text filter methods for the `db` source}
\value{
See [db-filter-discrete].
}
\description{
S7 methods implementing the `discrete_text` filter for a `dbtables` source.
They dispatch on `list(CbFilterDiscreteText, db_class)`.
}
