% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortBuilder.db-package.R
\name{db_class}
\alias{db_class}
\title{S7 class wrapper for the `db` source}
\format{
An S7 S3-class wrapper created with [S7::new_S3_class()].
}
\usage{
db_class
}
\description{
S7 representation of the `db` source class. Used as the source side of the
`cohortBuilder` S7 dual-dispatch filter generics when registering methods,
e.g. `S7::method(cb_filter_data, list(CbFilterDiscrete, db_class))`.
}
\keyword{internal}
