% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{.filter_position.db}
\alias{.filter_position.db}
\title{Create target filter position in UI panel for db source}
\usage{
\method{.filter_position}{db}(source, step_id, filter, ns, ...)
}
\arguments{
\item{source}{Source object of db type.}

\item{step_id}{Id of the step that the filter belongs to.}

\item{filter}{Filter object.}

\item{ns}{Namespace function.}

\item{...}{Extra arguments passed to related method.}
}
\value{
A CSS selector string locating the filter's dataset group.
}
\description{
Create target filter position in UI panel for db source
}
