% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_date_range.R, R/filter_discrete.R,
%   R/filter_discrete_text.R, R/filter_range.R
\name{cb_filter.date_range.db}
\alias{cb_filter.date_range.db}
\alias{cb_filter.discrete.db}
\alias{cb_filter.discrete_text.db}
\alias{cb_filter.range.db}
\title{Filters definition}
\usage{
\method{cb_filter.date_range}{db}(
  source,
  type = "date_range",
  id = .gen_id(),
  name = id,
  variable,
  range = NA,
  dataset,
  keep_na = TRUE,
  ...,
  description = NULL,
  active = TRUE
)

\method{cb_filter.discrete}{db}(
  source,
  type = "discrete",
  id = .gen_id(),
  name = id,
  variable,
  value = NA,
  dataset,
  keep_na = TRUE,
  ...,
  description = NULL,
  active = TRUE
)

\method{cb_filter.discrete_text}{db}(
  source,
  type = "discrete",
  id = .gen_id(),
  name = id,
  variable,
  value = NA,
  dataset,
  ...,
  description = NULL,
  active = TRUE
)

\method{cb_filter.range}{db}(
  source,
  type = "range",
  id = .gen_id(),
  name = id,
  variable,
  range = NA,
  dataset,
  keep_na = TRUE,
  ...,
  description = NULL,
  active = TRUE
)
}
\arguments{
\item{source}{Source object.}

\item{type}{Character string defining filter type (having class of the same value as type).}

\item{id}{Id of the filter.}

\item{name}{Filter name.}

\item{variable}{Dataset variable used for filtering.}

\item{range}{Variable range to be applied in filtering.}

\item{dataset}{Dataset name to be used for filtering.}

\item{keep_na}{If `TRUE`, NA values are included.}

\item{...}{Source type specific parameters (or extra ones if not matching specific S3 method arguments).}

\item{description}{Filter description (optional).}

\item{active}{If FALSE filter will be skipped during Cohort filtering.}

\item{value}{Value(s) to be used for filtering.}
}
\description{
Methods used for configuration of specific filter types for `dbtables` source.
}
