% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{.step_attrition.db}
\alias{.step_attrition.db}
\title{Generate output of attrition plot for db source}
\usage{
\method{.step_attrition}{db}(source, id, cohort, session, ...)
}
\arguments{
\item{source}{Source object of db type.}

\item{id}{Id of attrition output.}

\item{cohort}{Cohort object.}

\item{session}{Shiny session object.}

\item{...}{Extra arguments passed to specific method.}
}
\value{
List of two objects: `render` and `output` defining rendering and
    output placeholder for the step attrition plot feature.
}
\description{
Generate output of attrition plot for db source
}
