% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{autofilter.db}
\alias{autofilter.db}
\title{Generate filters definition based on the Source data}
\usage{
\method{autofilter}{db}(source, attach_as = c("step", "meta"), ...)
}
\arguments{
\item{source}{Source object.}

\item{attach_as}{Choose whether the filters should be attached as a new step,
or list of available filters (used in filtering panel when `new_step = "configure"`).
By default in \code{step}.}

\item{...}{Extra arguments passed to a specific method.}
}
\value{
Source object having step configuration attached.
}
\description{
The method should analyze source data structure, generate proper filters based on
the data (e.g. column types) and attach them to source.
}
