% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{set_source.db}
\alias{set_source.db}
\title{Create db-based Source}
\usage{
\method{set_source}{db}(
  dtconn,
  primary_keys = NULL,
  binding_keys = NULL,
  source_code = NULL,
  description = NULL,
  ...
)
}
\arguments{
\item{dtconn}{Object of class "db" defining connection to the database, schema and tables names.
Should be created with \link{dbtables}.}

\item{primary_keys}{Definition of data `primary_keys`. See \link[cohortBuilder]{primary_keys}.}

\item{binding_keys}{Binding keys definition see \link[cohortBuilder]{binding-keys}.}

\item{source_code}{An expression presenting low-level code for creating source.
When provided, used as a part of reproducible code output.}

\item{description}{A named list storing the source objects description.
Can be accessed with \link{description} Cohort method.}

\item{...}{Source type specific parameters. Available in `attributes` list of resulting object.}
}
\description{
This is a method for cohortBuilder [set_source()] generic.
It creates Source object which is based on connection to database tables.
}
