% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{source-layer}
\alias{source-layer}
\alias{.init_step.db}
\alias{.pre_filtering.db}
\alias{.collect_data.db}
\alias{.get_stats.db}
\title{Source compatibility methods for "db" layer.}
\usage{
\method{.init_step}{db}(source)

\method{.pre_filtering}{db}(source, data_object, step_id)

\method{.collect_data}{db}(source, data_object)

\method{.get_stats}{db}(source, data_object)
}
\arguments{
\item{source}{Source object of "db" type.}

\item{data_object}{Object that allows source data access.
For "db" source it's a list of connections to temporary tables in connected database.}

\item{step_id}{Id of the currently used step.}
}
\description{
These are methods for `cohortBuilder` generics used for integrity between source layers.
See \link[cohortBuilder]{source-layer}.
}
