% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{updating-data-statistics}
\alias{updating-data-statistics}
\alias{.update_data_stats.db}
\title{Update data statistics method for db source}
\usage{
\method{.update_data_stats}{db}(source, step_id, cohort, session)
}
\arguments{
\item{source}{Source object of db type.}

\item{step_id}{Id of the filtering step.}

\item{cohort}{Cohort object.}

\item{session}{Shiny session object.}
}
\description{
Update data statistics method for db source
}
